pyproject.toml
[tool.poetry]
name = "MLOps-HW"
version = "0.1.0"
description = "Homework for the MLOps course"
authors = ["KamillaZyal <85340058+KamillaZyal@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.1.4"
numpy = "^1.26.0"
pre-commit = "^3.6.0"
black = "^23.9.1"
flake8 = "^6.1.0"
lightning = "2.1"
isort = "^5.13.0"
torchvision = "^0.16.2"
omegaconf = "^2.3.0"
hydra-core = "^1.3.2"
scikit-learn = "^1.3.2"
tqdm = "v4.66.1"
dvc = {extras = ["gdrive"], version = "^3.38.1"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 90
target-version = ['py310']
include = '\.pyi?$'

[tool.isort]
profile = "hug"
py_version=310

[tool.flake8]
max-line-length = 90
exclude = [".git", "__pycache__"]
max-complexity = 10